requirements.txt
asgiref==3.9.1
Django==5.2.5
django-bootstrap5==25.2
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.3
pillow==11.2.1
whitenoise==6.9.0
gunicorn==23.0.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
marshmallow>=4.0.0
environs[django]==14.3.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
django-widget-tweaks==1.5.0
django-multiselectfield==1.0.1
